tech(api): move module validation schemas from tests to src
