+++ api/package.json
@@ -90,6 +90,7 @@
     "modulix:test": "npm run test:api:path -- 'tests/devcomp/unit/infrastructure/datasources/learning-content/module-datasource_test.js' 'tests/devcomp/acceptance/module-instantiation_test.js' 'tests/devcomp/unit/infrastructure/datasources/learning-content/validation/module-validation_test.js' 'tests/devcomp/integration/repositories/module-repository_test.js'"
   },
   "dependencies": {
+    "@1024pix/epreuves-components": "^2.8.0",
     "@1024pix/oppsy": "^1.0.2",
     "@aws-sdk/client-s3": "^3.121.0",
     "@aws-sdk/lib-storage": "^3.121.0",
@@ -119,6 +120,7 @@
     "file-type": "^21.0.0",
     "hapi-swagger": "^17.0.0",
     "hash-int": "^1.0.0",
+    "html-validate": "^10.0.0",
     "i18n": "^0.15.0",
     "iconv-lite": "^0.7.0",
     "ioredis": "5.8.0",
@@ -160,7 +162,6 @@
     "yargs": "^18.0.0"
   },
   "devDependencies": {
-    "@1024pix/epreuves-components": "^2.8.0",
     "@1024pix/eslint-plugin": "^2.1.17",
     "@eslint/compat": "^2.0.1",
     "@ls-lint/ls-lint": "^2.3.1",
@@ -176,7 +177,6 @@
     "eslint-plugin-n": "^17.23.2",
     "eslint-plugin-unicorn": "^62.0.0",
     "form-data": "^4.0.0",
-    "html-validate": "^10.0.0",
     "jsdoc-to-markdown": "^9.0.0",
     "knip": "^5.81.0",
     "mocha": "^11.7.5",
