+++ api/src/devcomp/infrastructure/datasources/learning-content/validation/element/custom-draft-element-schema.js
