+++ api/src/devcomp/infrastructure/datasources/learning-content/validation/element/custom-element-schema.js
