+++ api/src/devcomp/infrastructure/datasources/learning-content/validation/element/download-schema.js
