+++ api/src/devcomp/infrastructure/datasources/learning-content/validation/element/expand-schema.js
