+++ api/src/devcomp/infrastructure/datasources/learning-content/validation/element/image-schema.js
