+++ api/src/devcomp/infrastructure/datasources/learning-content/validation/element/proposal-content-schema.js
