+++ api/src/devcomp/infrastructure/datasources/learning-content/validation/element/qab-schema.js
