+++ api/src/devcomp/infrastructure/datasources/learning-content/validation/element/qcm-schema.js
