+++ api/src/devcomp/infrastructure/datasources/learning-content/validation/element/qcu-declarative-schema.js
