+++ api/src/devcomp/infrastructure/datasources/learning-content/validation/element/qcu-discovery-schema.js
