+++ api/src/devcomp/infrastructure/datasources/learning-content/validation/element/qrocm-schema.js
