+++ api/src/devcomp/infrastructure/datasources/learning-content/validation/element/short-video-schema.js
