+++ api/src/devcomp/infrastructure/datasources/learning-content/validation/element/text-schema.js
