+++ api/src/devcomp/infrastructure/datasources/learning-content/validation/module-schema.js
