+++ api/src/devcomp/infrastructure/datasources/learning-content/validation/utils.js
