+++ api/tests/devcomp/unit/infrastructure/datasources/learning-content/module-datasource_test.js
@@ -1,9 +1,9 @@
 import { ModuleDoesNotExistError } from '../../../../../../src/devcomp/domain/errors.js';
 import moduleDatasource from '../../../../../../src/devcomp/infrastructure/datasources/learning-content/module-datasource.js';
+import { moduleSchema } from '../../../../../../src/devcomp/infrastructure/datasources/learning-content/validation/module-schema.js';
 import { LearningContentResourceNotFound } from '../../../../../../src/shared/domain/errors.js';
 import { catchErr, expect } from '../../../../../test-helper.js';
 import { joiErrorParser } from './validation/joi-error-parser.js';
-import { moduleSchema } from './validation/module-schema.js';
 
 const modules = await moduleDatasource.list();
